+++ eNMS/models/scheduling.py
@@ -58,7 +58,7 @@ class Task(AbstractBase):
             app.scheduler.remove_job(self.aps_job_id)
         Session.commit()
 
-    def generate_row(self):
+    def generate_row(self, **kwargs):
         return super().generate_row() + [
             f"""
             <ul class="pagination pagination-lg" style="margin: 0px; width: 250px">
@@ -145,7 +145,7 @@ class Task(AbstractBase):
         Session.commit()
 
     def run_properties(self):
-        properties = {"payload": self.initial_payload, "task": self.id}
+        properties = {"task": self.id, **self.initial_payload}
         if self.devices:
             properties["devices"] = [device.id for device in self.devices]
         if self.pools:
@@ -232,7 +232,7 @@ class Event(AbstractBase):
     service = relationship("Service", back_populates="events")
     service_name = association_proxy("service", "name")
 
-    def generate_row(self):
+    def generate_row(self, **kwargs):
         return super().generate_row() + [
             f"""
             <ul class="pagination pagination-lg" style="margin: 0px; width: 150px">